Safari-609.1.11.1/LayoutTests/fast/forms/validationMessage-expected.txt

Test for validationMessage DOM property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


input patternMismatch: Match the requested format
input valueMissing: Fill out this field
textarea valueMissing: Fill out this field
select valueMissing: Select an item in the list
input typeMismatch: Enter an email address
input badInput: Enter a number
badInput and valueMissing:
PASS numberInput.validationMessage is nonRequiredBadInputMessage
PASS but.validationMessage is ''
PASS anoninput.validationMessage is ''
PASS happyFieldset.validationMessage is ''
PASS happySelect.validationMessage is ''
PASS happyOutput.validationMessage is ''
PASS happyObject.validationMessage is ''
PASS happyKeygen.validationMessage is ''
PASS successfullyParsed is true

TEST COMPLETE
